增加了 选择题，填空题，选项，填空 的增、删、改，查接口 修改了 Maperr

+++ back_end/src/main/java/com/example/back_end/service/impl/QuestionnaireServiceImpl.java
@@ -14,6 +14,22 @@ public class QuestionnaireServiceImpl implements QuestionnaireService {
     @Resource
     QuestionnaireMapper questionnaireMapper;
 
+    // 获取创建该问卷的 userId
+    @Override
+    public Integer findUserIdOfQuestionnaire(int questionnaireId) {
+        return questionnaireMapper.getUserIdOfQuestionnaire(questionnaireId);
+    }
+
+    // 检查问卷是否属于 account
+    @Override
+    public Boolean belongsToAccount(Account account, int questionnaireId) {
+        Integer userId = findUserIdOfQuestionnaire(questionnaireId);
+        if (userId == null || userId != account.getId()) { // 问卷不存在 或 问卷不属于 account
+            return false;
+        }
+        return true;
+    }
+
     // 创建问卷
     @Override
     public String createQuestionnaire(Account account, Questionnaire questionnaire) {
@@ -24,47 +40,42 @@ public class QuestionnaireServiceImpl implements QuestionnaireService {
         }
     }
 
+    // 删除问卷
     @Override
-    public String updateQuestionnaire(Questionnaire questionnaire) {
-        if (questionnaireMapper.updateQuestionnaire(questionnaire) > 0) {
-            return null;
-        } else {
-            return "未知错误，请联系管理员";
+    public String deleteQuestionnaire(Account account, int questionnaireId) {
+        if (belongsToAccount(account, questionnaireId)) { // 属于 account
+            if (questionnaireMapper.deleteQuestionnaire(questionnaireId) > 0) {
+                return null;
+            } else {
+                return "未知错误，删除失败，请联系管理员";
+            }
         }
+        return "没有权限";
     }
 
+    // 修改问卷
     @Override
-    public String deleteQuestionnaire(int questionnaireId) {
-        if (questionnaireMapper.deleteQuestionnaire(questionnaireId) > 0) {
-            return null;
-        } else {
-            return "未知错误，请联系管理员";
+    public String updateQuestionnaire(Account account, Questionnaire questionnaire) {
+        if (belongsToAccount(account, questionnaire.getQuestionnaireId())) { // 属于 account
+            if (questionnaireMapper.updateQuestionnaire(questionnaire) > 0) {
+                return null;
+            } else {
+                return "未知错误，修改失败，请联系管理员";
+            }
         }
+        return "没有权限";
     }
 
+    // 获取问卷
     @Override
-    public Questionnaire createrFindQuestionnaire(Account account, int questionnaireId) {
-        Questionnaire questionnaire =  questionnaireMapper.getQuestionnaire(questionnaireId);
-        if (questionnaire == null) {
-            return null;
+    public Questionnaire findQuestionnaire(Account account, int questionnaireId) {
+        if (belongsToAccount(account, questionnaireId)) { // 属于 account
+            return questionnaireMapper.getQuestionnaire(questionnaireId);
         }
-        Integer userId = questionnaireMapper.getUserIdOfQuestionnaire(questionnaireId);
-        if (userId != account.getId()) {// 不是创建者的问卷
-            return null;
-        }
-        return questionnaire;
+        return null;
     }
 
-    @Override
-    public Questionnaire getQuestionnaire(int questionnaireId) {
-        return questionnaireMapper.getQuestionnaire(questionnaireId);
-    }
-
-    @Override
-    public Integer userIdOfQuestionnaire(int questionnaireId) {
-        return questionnaireMapper.getUserIdOfQuestionnaire(questionnaireId);
-    }
-
+    // 获取该用户所有问卷
     @Override
     public List<Questionnaire> getAllQuestionnaire(Account account) {
         return questionnaireMapper.getQuestionnaires(account);
